fix pretrain
